Add full width ShmPutImage tests
Add tests for ShmPutImage that display the full width of the image,
instead of only displaying a sub-segment from a larger allocated
image. The tests added are ShmPutImageFullWidth and
AlignedShmPutImageFullWidth. These tests can help to track down
potential sub-optimal code paths for full width PutImage in the
Xorg fb layer.